create UserMenu component and integrate into main layout for cleaner structure and external control
