+++ src/pages/UserPage.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-// MUI components
+// MUI ui
 import { Container, Typography, Box, Button } from "@mui/material";
 
 // Import fetchMe and logout actions
